requirements.txt
aiogram
google-api-python-client
pytrends
python-dotenv
pytz
